api/requirements.txt
Flask==3.1.1
Flask-Moment==1.0.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
flask-cors==4.0.0